saving the posts default author on form submission using javascript. performance increase.
Instead of trying to modify the post's author during the save_post hook, change the post's author on form submission using javascript.

Improved the speed of adding a contributing author to a publication. rather than looping through all of the publications, just grab the first publication that has the selected pub code. Speed improvement when saving a posts.

https://app.asana.com/0/15421016095247/15421016095247